:truck: :horse: remove interfaces from chatbot store
