+++ libs/model/convs-mgr/conversations/admin/system/src/index.ts
@@ -1,3 +1,4 @@
 export * from './lib/platform-types.enum';
 
 export * from './lib/platform.interface';
+export * from './lib/event.interface';
