+++ libs/model/convs-mgr/conversations/admin/system/src/lib/event.interface.ts
@@ -1,3 +1,5 @@
+import { Block } from '@app/model/convs-mgr/conversations/chats';
+import { IObject } from '@iote/bricks';
 import { EventCategoryTypes } from './event-category-types.enum';
 
 export interface Event
@@ -8,4 +10,10 @@ export interface Event
   payload: {};
   origin: string; //Id of the party that triggered the event
   subject: string; //Id of the party affected by the event
+}
+
+
+export interface Activity extends IObject{
+  chatId: string;
+  block: Block;
 }
